% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{check_input_data}
\alias{check_input_data}
\title{check_input_data}
\usage{
check_input_data(arg_types, geoData = NULL, rtData = NULL)
}
\arguments{
\item{arg_types}{list, data types of arguments}

\item{geoData}{sf object, map data}

\item{rtData}{data.frame, rt estimates in the format {'Source':{'rtData':x, 'casesInfectionData':x, 'casesReportData':x, 'obsCasesData':x}, ...}}
}
\description{
Check input data
}
