% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data_ref.R
\name{default_data_ref}
\alias{default_data_ref}
\title{default_data_ref}
\usage{
default_data_ref()
}
\description{
Create a default reference for a input datasets. Also useful as a template when defining your own.
}
