% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_map_legend_ref.R
\name{default_map_legend_ref}
\alias{default_map_legend_ref}
\title{default_map_legend_ref}
\usage{
default_map_legend_ref()
}
\description{
Create a default reference for a map legend. Also useful as a template when defining your own.
}
