% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_ts_color_ref.R
\name{default_ts_color_ref}
\alias{default_ts_color_ref}
\title{default_ts_color_ref}
\usage{
default_ts_color_ref()
}
\description{
Create a default reference for time series colors. Also useful as a template when defining your own.
}
