% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_ts_colors.R
\name{default_ts_colors}
\alias{default_ts_colors}
\title{default_ts_colors}
\usage{
default_ts_colors()
}
\description{
Define default colors for time series plots

Can be overridden with a list of the same format
}
