% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubnationalUrls.R
\name{getSubnationalUrls}
\alias{getSubnationalUrls}
\title{Get Subnational Estimate Urls}
\usage{
getSubnationalUrls(path, areas)
}
\arguments{
\item{path}{A character string container the overall path to subnational estimates}

\item{areas}{A character vector listing the subnational estimates (assuming that listed in the
geoData with capitalisation and without capitalisation in the path).}
}
\value{
A named list of subnational urls.
}
\description{
Get Subnational Estimate Urls
}
\examples{

getSubnationalUrls(path = "https://epiforecasts.io/covid/posts/national/",
.                  areas = c('Afghanistan', 'Brazil', 'Colombia', 'India'))

}
