% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubregionalUrls.R
\name{getSubregionalUrls}
\alias{getSubregionalUrls}
\title{Get Subregional Estimate Urls}
\usage{
getSubregionalUrls(path, areas)
}
\arguments{
\item{path}{A character string container the overall path to subnational estimates}

\item{areas}{A character vector listing the subregional estimates (assuming that listed in the
geoData with capitalisation and without capitalisation in the path).}
}
\value{
A named list of subnational urls.
}
\description{
Get Subregional Estimate Urls
}
\examples{

getSubregionalUrls(path = "https://epiforecasts.io/covid/posts/national/",
.                  areas = c('Afghanistan', 'Brazil', 'Colombia', 'India'))

}
