% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_qualitative.R
\name{legend_qualitative}
\alias{legend_qualitative}
\title{legend_qualitative}
\usage{
legend_qualitative(variable_name, legend_values)
}
\arguments{
\item{variable_name}{string, name of the variable of this legend.}

\item{legend_values}{list, reference for legend colors in the format: {'value':'color', ...}.}
}
\description{
Create a qualitative legend.
}
