% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapWidget.R
\name{mapWidget}
\alias{mapWidget}
\title{mapWidget}
\usage{
mapWidget(
  geoData = NULL,
  rtData = NULL,
  width = 900,
  elementId = NULL,
  dryRun = FALSE
)
}
\arguments{
\item{geoData}{sf object, map data}

\item{rtData}{data.frame, rt estimates in the format {'Source':{'rtData':x, 'casesInfectionData':x, 'casesReportData':x, 'obsCasesData':x}, ...}}

\item{width}{integer, width in pixels}

\item{elementId}{string, id of element}

\item{dryRun}{Logical, defaults to FALSE. Should the function be tested without the widget being created.
Useful for checking the integrity of input data.}
}
\description{
Create a map of Rt summary data
}
