% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtD3.R
\name{RtD3}
\alias{RtD3}
\title{RtD3}
\usage{
RtD3(
  geoData = NULL,
  summaryData = NULL,
  rtData = NULL,
  casesInfectionData = NULL,
  casesReportData = NULL,
  obsCasesData = NULL,
  rtData_Deaths = NULL,
  casesInfectionData_Deaths = NULL,
  casesReportData_Deaths = NULL,
  activeArea = "United Kingdom",
  activeData = "R0",
  activeTime = "all",
  runDate = NULL,
  subregional_ref = NULL,
  width = 1000,
  elementId = NULL
)
}
\arguments{
\item{geoData}{sf object, map data}

\item{summaryData}{data.frame, summary data for mapping}

\item{rtData}{data.frame, rt estimates}

\item{casesInfectionData}{data.frame, cases by date of infection estimates}

\item{casesReportData}{data.frame, cases by date of report estimates}

\item{obsCasesData}{data.frame, observed cases data}

\item{rtData_Deaths}{data.frame, rt estimates (trained on deaths)}

\item{casesInfectionData_Deaths}{data.frame, cases by date of infection estimates (trained on deaths)}

\item{casesReportData_Deaths}{data.frame, cases by date of report estimates (trained on deaths)}

\item{activeArea}{character, the default area to plot (defaults to United Kingdom)}

\item{activeData}{character, the default dataset to plot (defaults to 'R0')}

\item{activeTime}{character, the default time window (defaults to 'all')}

\item{runDate}{character, date of estimate run in the format ('YYYY-MM-DD')}

\item{subregional_ref}{list, reference to subnational estimates in the format {'country_name':'url'}.}

\item{width}{integer, width in pixels}

\item{elementId}{string, id of element}
}
\description{
Create an Rt visualisation using D3
}
