% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWidget.R
\name{summaryWidget}
\alias{summaryWidget}
\title{summaryWidget}
\usage{
summaryWidget(
  geoData = NULL,
  rtData = NULL,
  data_ref = NULL,
  subregional_ref = NULL,
  ts_color_ref = NULL,
  ts_bar_color = "lightgrey",
  projection = "geoEquirectangular",
  map_legend_ref = NULL,
  credible_threshold = 10,
  width = NULL,
  activeArea = "United Kingdom",
  downloadUrl = NULL,
  dryRun = FALSE
)
}
\arguments{
\item{geoData}{sf object, map data}

\item{rtData}{data.frame, rt estimates in the format {'Source':{'rtData':x, 'casesInfectionData':x, 'casesReportData':x, 'obsCasesData':x}, ...}}

\item{data_ref}{list, reference for input data column names. Specify the column holding geometry to be symbolized {'rtData':{'geometry_name':'region'}, ...}}

\item{subregional_ref}{list, reference to subnational estimates in the format {'country_name':'url', ...}.}

\item{ts_color_ref}{list, reference for colors for time series plots.}

\item{ts_bar_color}{string, color of observed cases bars in time series plots.}

\item{projection}{string, map projection, must be named in \href{https://github.com/d3/d3-geo-projection#projections}{d3-geo-projection}.}

\item{map_legend_ref}{list, reference for map legend variables}

\item{credible_threshold}{integer, Threshold for credible intervals, maximum observed cases * this value will be removed.}

\item{width}{integer, Width of widget in pixels.}

\item{activeArea}{string, Area to symbolize first.}

\item{downloadUrl}{string, URL to download data.}

\item{dryRun}{Logical, defaults to FALSE. Should the function be tested without the widget being created.
Useful for checking the integrity of input data.}
}
\description{
Create an Rt visualisation using D3. Need convenience functions to define defaults
}
