% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWidget.R
\name{summaryWidget}
\alias{summaryWidget}
\title{summaryWidget}
\usage{
summaryWidget(
  geoData = NULL,
  summaryData = NULL,
  rtData = NULL,
  activeArea = "United Kingdom",
  activeTime = "all",
  runDate = NULL,
  subregional_ref = NULL,
  width = 900,
  elementId = NULL
)
}
\arguments{
\item{geoData}{sf object, map data}

\item{summaryData}{data.frame, summary data for mapping}

\item{rtData}{data.frame, rt estimates in the format {'Source':{'rtData':x, 'casesInfectionData':x, 'casesReportData':x, 'obsCasesData':x}, ...}}

\item{activeArea}{character, the default area to plot (defaults to United Kingdom)}

\item{activeTime}{character, the default time window (defaults to 'all')}

\item{runDate}{character, date of estimate run in the format ('YYYY-MM-DD')}

\item{subregional_ref}{list, reference to subnational estimates in the format {'country_name':'url', ...}.}

\item{width}{integer, width in pixels}

\item{elementId}{string, id of element}
}
\description{
Create an Rt visualisation using D3
}
